removeTicket function working
removeTicket() is now able to pull the correct ID from the table TR and
remove it from the .window element before running a reload() of the
database. The php back-end will need to be submitted with a ticket
number and user-name.

+ TO DO:
Update back-end to support accepting to arguments: ticket, and user.
Both can be submitted in a JSON string that gets decoded and used in
MySQL to update the background table.