removeTicket developed
TAC-METRO now supports the function to remove tickets via ajax call that
stringifies a JSON object and requests the ticket in removeTicket.php
via MySQL call to update the desired ticket number.

+ Works currently with the dbase_backup DB in MySQL. It will need to be
changed to the dbase DB for production. Login information can remain.

+ $.Dialog prompts the user for 'confirm' or 'deny' via button click,
and that sends the request over to removeTickets.php before the $.Dialog
window is closed and the DB refreshed after the ajax call.