Default Preferences Organized
Designated a more suitable color scheme for the default (when no options
are detected in LocalStorage) design of the page. Included a few
additional colors to make_opts.htm and plugged in the default options to
be checked onLoad.

TO DO:
+ Develop a $.Dialog or similar pop-up to appear when clicking Edit
Tickets shortcut.
+ Do same with click event for Weather live-tile.

+  Modify NewTicket function and RemoveTicket so they operate with the
correct Tickets database in MySQL back-end.

-- REMOVED TMP and TMP-THUMBS images as they are related to the
attachment app, not the TAC-METRO project.