Function Adjustmants
Made some light modifications to the TACLogin and PostTicket pages. The
function for opening a new ticket has been revised and will work in
Chrome and IE 10+ but needs further testing in Firefox due to [Object
object] errors in the response headers.

+ updated a few legacy pages looking for jQuery files that did not exist
any more. Updated some to include the latest jQuery library and did not
see any page-breaking bugs as a result.

-- Foundation needs to eventually be replaced in some ancillary pages,
but will be okay for now. There are a lot of associated files because of
this.