AutoLogin Completed
Working version and auto-signs in users with the localStorage value
populated in whichever browser is being used. Included legacy editions
of non-functional code.

Also updated and corrected SetOpts.JS to re-paint the site according to
preferences, whenever the database is refreshed. This ensures it will
automatically update every few minutes.